requirements.txt
flask==2.2.3
flask-smorest==0.42.0
flask-jwt-extended==4.5.2
flask-cors==4.0.0
marshmallow==3.19.0
pymongo==4.3.3
openai>=1.3.0
python-dotenv==1.0.0
gunicorn==20.1.0
bcrypt==4.0.1
regex==2023.8.8